# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Fast - API de gestion hotelera",
    "description": "API para la administracion integral de las operaciones del hotel.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/habitaciones": {
      "get": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Listar habitaciones",
        "description": "Lista las habitaciones de la gestion hotelera, con filtros opcionales por estado, tipo y limpieza. Devuelve siempre un arreglo completo.",
        "operationId": "listarHabitaciones",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EstadoHabitacion"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TipoHabitacion"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "limpieza",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EstadoLimpieza"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limpieza"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de habitaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitacionRead"
                  },
                  "title": "Response Listarhabitaciones"
                }
              }
            }
          },
          "422": {
            "description": "Filtro invalido (estado, tipo o limpieza)"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Rol sin permisos para consultar habitaciones"
          }
        }
      },
      "post": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Crear habitacion",
        "description": "Registra una habitacion nueva.",
        "operationId": "crearHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitacionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Habitacion creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionRead"
                }
              }
            }
          },
          "422": {
            "description": "Datos de habitacion invalidos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "409": {
            "description": "El numero de habitacion ya existe"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}": {
      "get": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Obtener habitacion",
        "description": "Obtiene una habitacion por su identificador.",
        "operationId": "obtenerHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Habitacion encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionRead"
                }
              }
            }
          },
          "422": {
            "description": "Solicitud invalida por datos o parametros incorrectos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Rol sin permisos para consultar habitaciones"
          },
          "404": {
            "description": "Habitacion no encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Actualizar habitacion",
        "description": "Actualiza los datos de una habitacion existente.",
        "operationId": "actualizarHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitacionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Habitacion actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionRead"
                }
              }
            }
          },
          "422": {
            "description": "Datos de habitacion invalidos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "404": {
            "description": "Habitacion no encontrada"
          },
          "409": {
            "description": "El numero de habitacion ya existe"
          }
        }
      },
      "delete": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Eliminar habitacion",
        "description": "Elimina una habitacion sin reservas futuras.",
        "operationId": "eliminarHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Habitacion eliminada"
          },
          "422": {
            "description": "Solicitud invalida por datos o parametros incorrectos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "404": {
            "description": "Habitacion no encontrada"
          },
          "409": {
            "description": "La habitacion tiene reservas futuras"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}/estado": {
      "patch": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Actualizar estado de habitacion",
        "description": "Cambia el estado operativo de una habitacion.",
        "operationId": "actualizarEstadoHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitacionEstado"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionRead"
                }
              }
            }
          },
          "422": {
            "description": "Estado de habitacion invalido"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Rol sin permisos para cambiar el estado"
          },
          "404": {
            "description": "Habitacion no encontrada"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}/limpieza": {
      "patch": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Actualizar limpieza de habitacion",
        "description": "Cambia el estado de limpieza de una habitacion. Se permite aunque la habitacion este OCUPADA.",
        "operationId": "actualizarLimpiezaHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitacionLimpieza"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Limpieza actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionRead"
                }
              }
            }
          },
          "422": {
            "description": "Estado de limpieza invalido"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Rol sin permisos para cambiar la limpieza"
          },
          "404": {
            "description": "Habitacion no encontrada"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}/imagenes": {
      "post": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Subir imagen de habitacion",
        "description": "Sube una imagen (JPEG, PNG o WEBP, maximo 5 MB) a una habitacion. Si es la primera imagen, queda marcada como principal.",
        "operationId": "subirImagenHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_subirImagenHabitacion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imagen subida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionImagenRead"
                }
              }
            }
          },
          "422": {
            "description": "Formato no permitido o archivo demasiado grande"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "404": {
            "description": "Habitacion no encontrada"
          },
          "409": {
            "description": "La habitacion ya alcanzo el maximo de imagenes"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}/imagenes/{imagen_id}": {
      "delete": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Eliminar imagen de habitacion",
        "description": "Elimina una imagen de la habitacion y su archivo. Si era la principal, promueve otra imagen como principal.",
        "operationId": "eliminarImagenHabitacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          },
          {
            "name": "imagen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Imagen Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Imagen eliminada"
          },
          "422": {
            "description": "Solicitud invalida por datos o parametros incorrectos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "404": {
            "description": "Habitacion o imagen no encontrada"
          }
        }
      }
    },
    "/api/habitaciones/{habitacion_id}/imagenes/{imagen_id}/principal": {
      "patch": {
        "tags": [
          "habitaciones"
        ],
        "summary": "Marcar imagen principal",
        "description": "Marca una imagen de la habitacion como principal.",
        "operationId": "marcarImagenPrincipal",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "habitacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Habitacion Id"
            }
          },
          {
            "name": "imagen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Imagen Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen marcada como principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitacionImagenRead"
                }
              }
            }
          },
          "422": {
            "description": "Solicitud invalida por datos o parametros incorrectos"
          },
          "401": {
            "description": "Token de acceso invalido o ausente"
          },
          "403": {
            "description": "Se requiere el rol administrador"
          },
          "404": {
            "description": "Habitacion o imagen no encontrada"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_subirImagenHabitacion": {
        "properties": {
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo",
            "description": "Archivo de imagen a subir"
          }
        },
        "type": "object",
        "required": [
          "archivo"
        ],
        "title": "Body_subirImagenHabitacion"
      },
...
      "EstadoHabitacion": {
        "type": "string",
        "enum": [
          "DISPONIBLE",
          "OCUPADA",
          "MANTENIMIENTO"
        ],
        "title": "EstadoHabitacion"
      },
      "EstadoLimpieza": {
        "type": "string",
        "enum": [
          "LIMPIA",
          "SUCIA"
        ],
        "title": "EstadoLimpieza"
      },
...
      "HabitacionCreate": {
        "properties": {
          "numero": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Numero"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoHabitacion"
          },
          "capacidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Capacidad"
          },
          "precioPorNoche": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Preciopornoche"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoHabitacion",
            "default": "DISPONIBLE"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "numero",
          "tipo",
          "capacidad",
          "precioPorNoche"
        ],
        "title": "HabitacionCreate"
      },
      "HabitacionEstado": {
        "properties": {
          "estado": {
            "$ref": "#/components/schemas/EstadoHabitacion"
          }
        },
        "type": "object",
        "required": [
          "estado"
        ],
        "title": "HabitacionEstado"
      },
      "HabitacionImagenRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "orden": {
            "type": "integer",
            "title": "Orden"
          },
          "esPrincipal": {
            "type": "boolean",
            "title": "Esprincipal"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "orden",
          "esPrincipal",
          "url"
        ],
        "title": "HabitacionImagenRead"
      },
      "HabitacionLimpieza": {
        "properties": {
          "limpieza": {
            "$ref": "#/components/schemas/EstadoLimpieza"
          }
        },
        "type": "object",
        "required": [
          "limpieza"
        ],
        "title": "HabitacionLimpieza"
      },
      "HabitacionRead": {
        "properties": {
          "numero": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Numero"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoHabitacion"
          },
          "capacidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Capacidad"
          },
          "precioPorNoche": {
            "type": "number",
            "title": "Preciopornoche"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoHabitacion",
            "default": "DISPONIBLE"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "limpieza": {
            "$ref": "#/components/schemas/EstadoLimpieza"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "imagenes": {
            "items": {
              "$ref": "#/components/schemas/HabitacionImagenRead"
            },
            "type": "array",
            "title": "Imagenes",
            "default": []
          },
          "imagenPrincipalUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagenprincipalurl",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "numero",
          "tipo",
          "capacidad",
          "precioPorNoche",
          "id",
          "limpieza",
          "createdAt",
          "updatedAt",
          "imagenPrincipalUrl"
        ],
        "title": "HabitacionRead"
      },
...
      "HabitacionUpdate": {
        "properties": {
          "numero": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Numero"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoHabitacion"
          },
          "capacidad": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Capacidad"
          },
          "precioPorNoche": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Preciopornoche"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoHabitacion",
            "default": "DISPONIBLE"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "numero",
          "tipo",
          "capacidad",
          "precioPorNoche"
        ],
        "title": "HabitacionUpdate"
      },
...
      "TipoHabitacion": {
        "type": "string",
        "enum": [
          "SIMPLE",
          "DOBLE",
          "SUITE",
          "PRESIDENCIAL"
        ],
        "title": "TipoHabitacion"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}